Move `Action` to the package level
